+++ tests/Classifiers/ExtraTreeClassifierTest.php
@@ -7,10 +7,9 @@ use Rubix\ML\DataType;
 use Rubix\ML\Estimator;
 use Rubix\ML\Persistable;
 use Rubix\ML\Probabilistic;
+use Rubix\ML\RanksFeatures;
 use Rubix\ML\EstimatorType;
 use Rubix\ML\Datasets\Unlabeled;
-use Rubix\ML\Graph\Trees\ExtraTree;
-use Rubix\ML\Graph\Trees\DecisionTree;
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\Classifiers\ExtraTreeClassifier;
 use Rubix\ML\Datasets\Generators\Agglomerate;
@@ -97,11 +96,10 @@ class ExtraTreeClassifierTest extends TestCase
     public function build() : void
     {
         $this->assertInstanceOf(ExtraTreeClassifier::class, $this->estimator);
-        $this->assertInstanceOf(ExtraTree::class, $this->estimator);
-        $this->assertInstanceOf(DecisionTree::class, $this->estimator);
+        $this->assertInstanceOf(Estimator::class, $this->estimator);
         $this->assertInstanceOf(Learner::class, $this->estimator);
         $this->assertInstanceOf(Probabilistic::class, $this->estimator);
-        $this->assertInstanceOf(Estimator::class, $this->estimator);
+        $this->assertInstanceOf(RanksFeatures::class, $this->estimator);
         $this->assertInstanceOf(Persistable::class, $this->estimator);
     }